auth

+++ src/services/contacts.js
@@ -1,5 +1,5 @@
 import { SORT_ORDER } from "../constants/index.js";
-import ContactCollection from "../db/models/Contact.js";
+import ContactCollection from '../db/models/Contact.js';
 import { calculatePaginationData } from "../utils/calculatePagenationData.js";
 
 export const getContacts = async ({
@@ -8,11 +8,12 @@ export const getContacts = async ({
     sortOrder = SORT_ORDER.ASC,
     sortBy = '_id',
     filter = {},
+    userId,
 }) => {
     const limit = perPage;
     const skip = (page - 1) * perPage;
 
-    const contactsQuery = ContactCollection.find();
+    const contactsQuery = ContactCollection.find({userId});
 
     if (filter.contactType) {
         contactsQuery.where('contactType').regex(new RegExp(filter.contactType, 'i'));
@@ -26,7 +27,7 @@ export const getContacts = async ({
     // const contacts = await contactsQuery.skip(skip).limit(limit).sort({[sortBy]: sortOrder}).exec();
 
     const [contactsCount, contacts] = await Promise.all([
-        ContactCollection.find().merge(contactsQuery).countDocuments(),
+        ContactCollection.find({userId}).merge(contactsQuery).countDocuments(),
         contactsQuery
             .skip(skip)
             .limit(limit)
@@ -41,21 +42,21 @@ export const getContacts = async ({
     };
 };
 
-export const getContactById = contactId => ContactCollection.findOne({_id: contactId});
+export const getContactById = (contactId, userId) => ContactCollection.findOne({_id: contactId, userId});
 
-export const createContact = async (payload) => {
+export const createContact = async (payload, userId) => {
     const contact = await ContactCollection.create(payload);  
     return contact;  
 };
 
-export const deleteContact = async (contactId) => {
-    const contact = await ContactCollection.findOneAndDelete({_id: contactId});
+export const deleteContact = async (contactId, userId) => {
+    const contact = await ContactCollection.findOneAndDelete({_id: contactId, userId});
     return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
+export const updateContact = async (contactId, payload, userId, options = {}) => {
     const contact = await ContactCollection.findOneAndUpdate(
-        {_id: contactId}, 
+        {_id: contactId, userId}, 
         payload,
         {
             new: true,
